# server/cates/cates.create_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 15:51:44 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 15:51:44 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package cates

import (
	"github.com/bachhieu/fountain/baselib/base"
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/bachhieu/test/biz/dal/do/cate_do"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Create Cate
// @Description  Create a new Cate
// @Tags         Cate
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string   true  "Insert your access token" default(access_token)
// @Param        request       body      cate_do.CreateCateReq  true  "Create Cate request"
// @Success      200          {object}  v_api.Response[cate_do.CateResponse]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/cates [post]
func (api *CatesAPI) Create(ctx *fiber.Ctx) error {
	input := &cate_do.CreateCateReq{}

	if err := ctx.BodyParser(input); err != nil {
		err := v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), err.Error())
		v_log.V(1).WithError(err).Errorf("CatesAPI::Create - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}

	input.CreatedBy = v_api.GetContextDataString(ctx, v_api.KContextKeyUserID)
	if err := api.catesController.Create(input); err != nil {
		v_log.V(1).WithError(err).Errorf("CatesAPI::Create - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}

	v_log.V(3).Infof("CatesAPI::Create - Reply: %s", base.JSONDebugDataString(input))

	return v_api.WriteSuccess(ctx, input)
}