# server/cates/cates.get_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 15:51:44 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 15:51:44 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package cates

import (
	"github.com/bachhieu/fountain/baselib/base"
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/bachhieu/test/biz/dal/do/cate_do"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Get Cate
// @Description  Get list of Cate with optional filters
// @Tags         Cate
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true   "Insert your access token" default(access_token)
// @Param        id     query     string  true  "ID of the Cate"
// @Success      200          {object}  v_api.Response[cate_do.CateResponse]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/cates/detail [get]
func (api *CatesAPI) Get(ctx *fiber.Ctx) error {
	id := ctx.Query("id", "")
	if id == "" {
		return v_api.WriteError(ctx, v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "id is required"))
	}
	res := api.catesController.Get(id)

	v_log.V(3).Infof("CatesAPI::Get - Reply: %s", base.JSONDebugDataString(res))

	return v_api.WriteSuccess(ctx, cate_do.NewCateResponse(res))
}