# server/cates/cates.pagination_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 15:51:44 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 15:51:44 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package cates

import (
	"github.com/bachhieu/fountain/baselib/base"
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/bachhieu/test/biz/dal/do/cate_do"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Get Pagination
// @Description  Get pagination of Cate
// @Tags         Cate
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true  "Insert your access token" default(access_token)
// @Param        query_params              query     cate_do.CateQueryParams  false  "Query params"
// @Success      200          {object}  v_api.Response[[]cate_do.CateResponse]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/cates/pagination [get]
func (api *CatesAPI) GetPagination(ctx *fiber.Ctx) error {
	params := new(cate_do.CateQueryParams)
	if err := ctx.QueryParser(params); err != nil {
		v_log.V(1).WithError(err).Errorf("CatesAPI::GetPagination - Parse params error: %+v", err)
		return v_api.WriteError(ctx, v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "invalid parameters"))
	}

	offset, limit := v_api.GetQueryOffsetLimit(ctx)
	res, err := api.catesController.Gets(params, offset, limit)
	if err != nil {
		v_log.V(1).WithError(err).Errorf("CatesAPI::GetPagination - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}

	total := api.catesController.CountTotal(params)

	v_log.V(3).Infof("CatesAPI::GetPagination - Reply: %s", base.JSONDebugDataString(res))

	return v_api.WithContext(ctx).WithPaginationByLoadMore(total, offset, limit).WriteSuccess(cate_do.NewCateResponses(res))
}