# server/products/products.delete_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 16:10:50 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 16:10:50 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package products

import (
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Delete Product
// @Description  Delete an existing Product
// @Tags         Product
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true  "Insert your access token" default(access_token)
// @Param        id     query     string  true  "Product ID"
// @Success      200          {object}  v_api.Response[string]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/products [delete]
func (api *ProductsAPI) Delete(ctx *fiber.Ctx) error {
	id := ctx.Query("id")
	if id == "" {
		err := v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "id is required")
		v_log.V(1).WithError(err).Errorf("ProductsAPI::Delete - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}
	if err := api.productsController.Delete(id); err != nil {
		v_log.V(1).WithError(err).Errorf("ProductsAPI::Delete - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}

	v_log.V(3).Infof("ProductsAPI::Delete - Reply: Oke")

	return v_api.WriteSuccess(ctx, "Deleted successfully")
}