# server/products/products.get_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 16:10:50 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 16:10:50 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package products

import (
	"github.com/bachhieu/fountain/baselib/base"
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/bachhieu/test/biz/dal/do/product_do"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Get Product
// @Description  Get list of Product with optional filters
// @Tags         Product
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true   "Insert your access token" default(access_token)
// @Param        id     query     string  true  "ID of the Product"
// @Success      200          {object}  v_api.Response[product_do.ProductResponse]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/products/detail [get]
func (api *ProductsAPI) Get(ctx *fiber.Ctx) error {
	id := ctx.Query("id", "")
	if id == "" {
		return v_api.WriteError(ctx, v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "id is required"))
	}
	res := api.productsController.Get(id)

	v_log.V(3).Infof("ProductsAPI::Get - Reply: %s", base.JSONDebugDataString(res))

	return v_api.WriteSuccess(ctx, product_do.NewProductResponse(res))
}